Move patcher to _perf directory
It's not perf specific, but perf is the only one using it so keep it hidden
for now, we can always move it later.
