+++ napari/utils/perf/_config.py
@@ -8,8 +8,8 @@ from typing import List, Optional
 
 import wrapt
 
-from ...utils.patcher import patch_callables
 from ._utils import perf_timer
+from ._patcher import patch_callables
 
 PERFMON_ENV_VAR = "NAPARI_PERFMON"
