+++ napari/utils/perf/_patcher.py
